+++ src/main.tsx
@@ -23,10 +23,11 @@ import Art from "./pages/Admin/MyArt";
 import CreateArt from "./pages/Admin/CreateArt";
 
 import { ADMIN_ROUTE, SIGNUP_ROUTE } from "./constants/router";
-import GalleryHOC from "./pages/Home/Gallery/GalleryHOC";
+import GalleryHOC from "./pages/Home/Gallery/Gallery";
 import SearchGallery from "./pages/Search";
 import { RecoilRoot } from "recoil";
 import Signup from "./pages/Signup";
+import Gallery from "./pages/Home/Gallery/Gallery";
 
 const router = createBrowserRouter([
   {
@@ -42,7 +43,7 @@ const router = createBrowserRouter([
         children: [
           {
             index: true,
-            element: <GalleryHOC />
+            element: <Gallery />
           },
           {
             path: "search",
