+++ src/pages/Home/Gallery/GalleryHOC.tsx
@@ -1,28 +1,44 @@
-import { useRecoilValue } from "recoil";
-import Gallery from ".";
-import useGetThumbnails from "../../../hooks/Gallery/useGetThumbnails";
-import { galleryArts } from "../../../store/gallery";
 import { useEffect } from "react";
 import { TGalleryProps } from "./types";
+import GalleryView from "./View";
+import { useInView } from "react-intersection-observer";
 
-const GalleryHOC = () => {
-  const thumbnails = useRecoilValue(galleryArts);
+const GalleryHOC = ({
+  thumbnails,
+  data,
+  setNextQuery,
+  isLoading
+}: TGalleryProps) => {
+  const { ref: intObserver, inView } = useInView({ threshold: 0.8 });
 
-  const { data, isLoading, setNextQuery, refetch } = useGetThumbnails();
-  const GalleryHOCProps: TGalleryProps = {
-    thumbnails,
-    data,
-    isLoading,
-    setNextQuery
+  const isScrolledToBottom = () => {
+    const { scrollTop, scrollHeight, clientHeight } = document.documentElement;
+    return scrollTop + clientHeight >= scrollHeight - 100;
+  };
+
+  const readyToFetchNext = isScrolledToBottom() && inView;
+
+  const onIntersection = () => {
+    if (data?.hasNext)
+      setNextQuery({
+        date: data.nextRevisionDateIdx,
+        idx: data.nextIdx
+      });
   };
 
   useEffect(() => {
-    if (thumbnails.length === 0) {
-      refetch();
-    }
-  }, []);
+    if (!readyToFetchNext) return;
+    onIntersection();
+  }, [readyToFetchNext]);
 
-  return <Gallery {...GalleryHOCProps} />;
+  const GalleryProps: TGalleryProps = {
+    thumbnails,
+    isLoading,
+    hasNext: data?.hasNext,
+    setNextQuery
+  };
+
+  return <GalleryView {...GalleryProps} ref={intObserver} />;
 };
 
 export default GalleryHOC;
